web/2025hack/2025hack/requirements.txt
numpy==1.23.5
opencv-python==4.6.0.66
tensorflow==2.11.0
pillow==9.3.0
tqdm==4.64.1
gin-config
scipy==1.9.3
mediapipe==0.10.5